+++ artifacts/omniproject/src/pages/Settings.tsx
@@ -31,6 +31,7 @@ import { IdentityMapAdmin } from "../components/settings/IdentityMapAdmin";
 import { CostRulesAdmin } from "../components/settings/CostRulesAdmin";
 import { CustomReportsAdmin } from "../components/settings/CustomReportsAdmin";
 import { ContentPagesAdmin } from "../components/settings/ContentPagesAdmin";
+import { PriorityWeightsAdmin } from "../components/settings/PriorityWeightsAdmin";
 import { GovernanceRulesAdmin } from "../components/settings/GovernanceRulesAdmin";
 import { ScopeUpliftAdmin } from "../components/settings/ScopeUpliftAdmin";
 import { FeatureGovernance } from "../components/settings/FeatureGovernance";
@@ -380,6 +381,10 @@ export function Settings() {
 
       <div className="mt-10">
         <ContentPagesAdmin />
+      </div>
+
+      <div className="mt-10">
+        <PriorityWeightsAdmin />
       </div>
 
       <div className="mt-10">